requirements.txt
numpy
pandas
mecab-python3
ipadic
sklearn
scipy
matplotlib
jupyterlab
